Add files via upload
This Jupyter Notebook contains a detailed analysis and implementation for detecting liver disease using machine learning techniques. The notebook begins with data preprocessing, where steps are taken to clean and prepare the dataset for modeling. Following this, an exploratory data analysis (EDA) is conducted, providing visualizations and statistical summaries to understand the distribution and relationships within the data.

Feature engineering is then applied, using various techniques to enhance the dataset and improve model performance. The notebook proceeds with the implementation of different machine learning algorithms to predict liver disease, including model training, evaluation, and comparison. Additionally, hyperparameter tuning is performed to optimize the model parameters and improve accuracy.